sampleAPI_basic/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
hupper==1.10.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
Werkzeug==2.0.1
